fix(snapshot): fix snap usage
Closes #40
