+++ src/index.js
@@ -1,6 +1,7 @@
 import * as tracker from './componentTracker'
 
 export { default as loadComponents } from './loadComponents'
+export { default as getState } from './getState'
 export { LOADABLE } from './constants'
 export { default } from './loadable'
 export const componentTracker = tracker
